マップ描画

+++ app/is_exercise/CanalWishes/src/main/java/com/moai/cw/game_object/Block.java
@@ -14,7 +14,7 @@ public class Block extends Entity {
   }
 
   @Override
-  public void update(double dt) {
+  public void update(int dt) {
 
   }
 }
