+++ app/is_exercise/CanalWishes/src/main/java/com/moai/cw/scene/FieldScene.java
@@ -16,7 +16,7 @@ public class FieldScene extends Scene {
   }
 
   @Override
-  public void update(double dt) {
+  public void update(int dt) {
 
   }
 }
